Header: restyling, add profile image
