+++ src/components/RightPanel.js
@@ -15,9 +15,9 @@ import MusicList, {
     TRANSFORM_RECENT,
     TRANSFORM_RECENT_SINGLE
 } from "./MusicList";
-import CustomTag from "./CustomTag";
 import TagList from "./TagList";
 import { useHistory } from "react-router";
+import SingleCollaborator from "./SingleCollaborator";
 
 //const DEFAULT_BACKGROUND = 'https://www.rollingstone.com/wp-content/uploads/2018/09/beatles-white-album-.jpg';
 
@@ -32,17 +32,6 @@ function getCollaborators(project) {
         }
     }
     return collaborators;
-}
-
-function SingleCollaborator(props) {
-    return (
-        <div className="flex flex-col m-4">
-            <div className="rounded-full h-24 w-24 overflow-hidden">
-                <img src={props.profileImage} className="object-cover h-24" />
-            </div>
-            <p className="text-center"> {props.name} </p>
-        </div>
-    );
 }
 
 function AllCollaborators(props) {
