+++ src/hocs/withHeader/Header.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useEffect} from "react";
 import { Link } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
 import Logo from "./logo.svg";
@@ -45,19 +45,24 @@ const Header = () => {
                     to="/"
                     className="flex flex-row items-center justify-start"
                 >
-                    {/* <img src={Logo} alt="logo" /> */}
-                    <div className="w-10 h-10 shadow-inner box-border rounded-full bg-red-600 mr-4" />
-                    <h1 className="flex-1"> SoundBeat </h1>
+                    <img src={"https://icon-library.com/images/analyze-_sound-_wave-music-512_362.png"} alt="logo" className="h-24 p-2"/>
+                    {/* <div className="w-10 h-10 shadow-inner box-border rounded-full bg-red-600 mr-4" /> */}
+                    <h1 className="flex-1 text-indigo-600"> MusicCollab </h1>
                 </Link>
             </div>
             <MusicPlayer />
 
             <div className="flex flex-1 flex-row items-center justify-end">
-                <h1> {userName} </h1>
                 {!userId ? (
                     <GenericButton title={"Login"} onClick={onLogin} />
                 ) : (
                     <>
+                        <div className="flex flex-row items-center">
+                            <div className="rounded-full h-12 w-12 overflow-hidden">
+                                <img src={profiles[userId].metaInfo.profileImage} className="object-cover h-12" />
+                            </div>
+                            <h1 className="pl-4 text-3xl"> {userName} </h1>
+                        </div>
                         <div
                             data-cy="separator"
                             className="w-16 h-0 transform rotate-90 border-2 border-gray-20"
